"Efetuada logica de verificação do cvc onde nao é possivel encerrar a compra sem confirmar o mesmo e implementado a refatoração no banco para amarrar as compras com o endereço de entraga e os dados de pagamento."

+++ lojavinho/src/main/java/br/com/lojavinho/servlet/EncerrarCompraServlet.java
@@ -2,13 +2,14 @@ package br.com.lojavinho.servlet;
 
 import br.com.lojavinho.dao.CarrinhoDao;
 import br.com.lojavinho.dao.ComprasDao;
-import br.com.lojavinho.model.Compras;
-import br.com.lojavinho.model.ItemCarrinho;
+import br.com.lojavinho.dao.PagamentoDao;
+import br.com.lojavinho.model.*;
 
 import java.io.IOException;
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
+import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
@@ -19,26 +20,82 @@ import javax.servlet.http.HttpSession;
 @WebServlet("/EncerrarCompra")
 public class EncerrarCompraServlet extends HttpServlet {
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        String numCPF = request.getParameter("numCPF");
-        String numSeqPag = request.getParameter("numSeqPag");
-        String numSeqEntrega = request.getParameter("numSeqEntrega");
+        HttpSession session = request.getSession();
+        String numCPF = (String) session.getAttribute("CPFCliente");
 
-        List<ItemCarrinho> listaCarrinho = CarrinhoDao.lerItemCarrinho(numCPF);
+        String cvcParam = request.getParameter("cvc");
 
-        BigDecimal valorTotalVenda = calcularValorTotal(listaCarrinho);
+        if (cvcParam == null || cvcParam.trim().isEmpty()) {
+            DadosEntrega dadosEntrega = ComprasDao.obterUltimaCompraPorCPF(numCPF);
+            request.setAttribute("numCPF", dadosEntrega.getNumCPF());
+            request.setAttribute("CEP", dadosEntrega.getCEP());
+            request.setAttribute("endereco", dadosEntrega.getEndereco());
+            request.setAttribute("numEndereco", dadosEntrega.getNumEndereco());
+            request.setAttribute("complEndereco", dadosEntrega.getComplEndereco());
+            request.setAttribute("bairro", dadosEntrega.getBairro());
+            request.setAttribute("cidade", dadosEntrega.getCidade());
+            request.setAttribute("estado", dadosEntrega.getEstado());
 
-        Compras compras = new Compras();
-        compras.setCpfCliente(numCPF);
-        compras.setDataOperacao(LocalDateTime.now());
-        compras.setValorTotalVenda(valorTotalVenda);
-        compras.setNumSeqPag(numSeqPag);
-        compras.setNumSeqEntrega(numSeqEntrega);
+            DadosPagamento dadosPagamento = ComprasDao.obterDadosPagamentoPorCPF(numCPF);
+            request.setAttribute("titularCartao", dadosPagamento.getNomeCartao());
+            request.setAttribute("dataValidade", dadosPagamento.getDataValidade());
+            request.setAttribute("numCartao", dadosPagamento.getNumCartao());
 
-        ComprasDao.registrarCompra(compras, listaCarrinho);
+            String qtdTotal = request.getParameter("qtdTotal");
+            String vlrTotal = request.getParameter("vlrTotal");
 
-        CarrinhoDao.excluirCarrinho(numCPF);
+            Carrinho carrinho = new Carrinho(numCPF, qtdTotal, vlrTotal);
 
-        response.sendRedirect("/PaginaDeConfirmacao/paginaDeConfirmacao.jsp");
+            request.setAttribute("carrinho", carrinho);
+
+            List<ItemCarrinho> listaCarrinho = CarrinhoDao.lerItemCarrinho(numCPF);
+            request.setAttribute("listaCarrinho", listaCarrinho);
+
+            String erroCvc = "Forneça um CVC válido para finalizar a compra!.";
+            request.setAttribute("erroCvc", erroCvc);
+
+            RequestDispatcher dispatcher = request.getRequestDispatcher("/TelaFinalizarCompra/FinalizarCompra.jsp");
+            dispatcher.forward(request, response);
+        } else {
+            try {
+                int cvc = Integer.parseInt(cvcParam);
+
+                boolean cvcExiste = PagamentoDao.verificarCvc(numCPF, String.valueOf(cvc));
+
+                if (cvcExiste) {
+                    List<ItemCarrinho> listaCarrinho = CarrinhoDao.lerItemCarrinho(numCPF);
+                    numCPF = request.getParameter("numCPF");
+                    String numSeqPag = request.getParameter("numSeqPag");
+                    String numSeqEntrega = request.getParameter("numSeqEntrega");
+
+
+                    BigDecimal valorTotalVenda = calcularValorTotal(listaCarrinho);
+
+                    Compras compras = new Compras();
+                    compras.setCpfCliente(numCPF);
+                    compras.setDataOperacao(LocalDateTime.now());
+                    compras.setValorTotalVenda(valorTotalVenda);
+                    compras.setNumSeqPag(numSeqPag);
+                    compras.setNumSeqEntrega(numSeqEntrega);
+
+                    ComprasDao.registrarCompra(compras, listaCarrinho);
+
+                    CarrinhoDao.excluirCarrinho(numCPF);
+
+                    response.sendRedirect("/PaginaDeConfirmacao/paginaDeConfirmacao.jsp");
+                } else {
+
+                    request.setAttribute("erroCvc", "Código CVC incorreto. Por favor, verifique e tente novamente.");
+                    RequestDispatcher dispatcher = request.getRequestDispatcher("/TelaFinalizarCompra/FinalizarCompra.jsp");
+                    dispatcher.forward(request, response);
+                }
+            } catch (NumberFormatException e) {
+
+                request.setAttribute("erroCvc", "Por favor, forneça um código CVC válido.");
+                RequestDispatcher dispatcher = request.getRequestDispatcher("/TelaCadastroPagamento/CadastroPagamento.jsp");
+                dispatcher.forward(request, response);
+            }
+        }
     }
 
     private BigDecimal calcularValorTotal(List<ItemCarrinho> listaCarrinho) {
